+++ JrgbCore/src/main/java/com/github/nh0815/jrgb/rgb/IdentityRGBStrategy.java
@@ -1,4 +1,4 @@
-package com.github.nh0815.rgb;
+package com.github.nh0815.jrgb.rgb;
 
 /**
  * Created by Nick on 10/23/2015.
