+++ JrgbUI/src/main/java/com/github/nh0815/jrgb/ui/RgbPanel.java
@@ -1,7 +1,6 @@
 package com.github.nh0815.jrgb.ui;
 
-import com.github.nh0815.RgbConverter;
-import com.github.nh0815.rgb.RGBStrategy;
+import com.github.nh0815.jrgb.RgbConverter;
 
 import javax.swing.*;
 import java.awt.*;
