+++ JrgbUI/src/main/java/com/github/nh0815/jrgb/ui/RgbWindow.java
@@ -1,7 +1,7 @@
 package com.github.nh0815.jrgb.ui;
 
-import com.github.nh0815.RgbConverter;
-import com.github.nh0815.rgb.*;
+import com.github.nh0815.jrgb.RgbConverter;
+import com.github.nh0815.jrgb.rgb.*;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
